BUGS: 'hasattr("summary")' yields infinite loop in py3 and squeeze Y
Something about the `property` decorator has changed in py3 that causes
an infinite loop with how I had it set up before. It should be fixed
now.

The regression functions were supposed to be able to take a string for
`y_name` or a list of length 1. However, the list didn't actually work
because the y object wouldb be a DataFrame instead of a Series. The y
object now gets squeezed in case it is a DataFrame with 1 column.